+++ R/expressions.R
@@ -59,7 +59,7 @@ math.exp <- function (e1, e2, operator) {
         ## Because of how this function is invoked, get the offending expression
         ## from the call before this one
         halt("Invalid expression (probably a reference to a variable that doesn't exist): ",
-            deparse(tail(sys.calls(), 2)[[1]]))
+            deparseAndTruncate(tail(sys.calls(), 2)[[1]]))
     }
     ex <- zfunc(operator, e1, e2)
     ds.url <- unique(unlist(lapply(list(e1, e2), datasetReference))) %||% ""
@@ -305,7 +305,7 @@ setMethod("[", c("CrunchExpr", "CrunchLogicalExpr"), .updateActiveFilter)
         ## If you reference a variable in a dataset that doesn't exist, you
         ## get NULL, and e.g. NULL == something becomes logical(0).
         ## That does awful things if you try to send to the server. So don't.
-        halt("Invalid expression: ", deparse(match.call()$i)[1])
+        halt("Invalid expression: ", deparseAndTruncate(match.call()$i))
     }
 }