+++ tests/testthat/test-filters.R
@@ -98,6 +98,13 @@ with_mock_HTTP({
             '{"variable":"https://app.crunch.io/api/datasets/1/variables/gender/"},',
             '{"value":1}]}}')
     })
+
+    test_that("Print method for filter entity (debug)", {
+        f <- CrunchFilter(crGET("a-filter/"))
+        expect_is(f, "CrunchFilter")
+        expect_fixed_output(f,
+            'starttime %in% c("2016-04-06", "2016-04-15", "2016-04-25", "2016-05-06", "2016-05-13", ... & gender %in% "Male"')
+    })
 })
 
 with_test_authentication({